Add support for tagging and querying tags.
The `tag` and `query` commands now work! The default state is to tag the
currently playing song on spotify, which is a super useful way of making
it easy to just throw tags onto stuff.

Lots of potenital next steps (e.g. adding ratings and ordering by them
in the output seems like a useful one!), but for now this is a solid
mvp.